# services/gateway-svc/api/rest_oapi/openapi.yaml
openapi: 3.0.3
info:
  title: gateway-svc API
  version: 0.1.0
  description: |
    Edge REST API for UmrohOS. gateway-svc fronts every backend service and is
    the only outward-facing HTTP surface clients (web app, mobile app, CLI)
    should call.

    Scaffold scope: 12 endpoints —

      - 2 own probes: /system/live, /system/ready
      - 10 per-backend liveness proxies: /v1/<svc>/system/live, one per
        backend in the umrohos service map. Used by the web app's status page.

    Real per-route proxies (auth, packages, bookings, ...) land alongside
    each backend's first feature work.
servers:
  - url: /
    description: Base path
paths:
...
  /v1/packages:
    get:
      tags: [Catalog]
      summary: List active packages
      operationId: listPackages
      security: []
      parameters:
        - name: kind
          in: query
          required: false
          schema: { $ref: "#/components/schemas/PackageKind" }
        - name: airline_code
          in: query
          required: false
          schema: { type: string }
        - name: hotel_id
          in: query
          required: false
          schema: { type: string }
        - name: departure_from
          in: query
          required: false
          schema: { type: string, format: date }
        - name: departure_to
          in: query
          required: false
          schema: { type: string, format: date }
        - name: cursor
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListPackagesResponse" }
        "400":
          description: Bad request (invalid_query_param / invalid_cursor)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "500":
          description: Internal error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "502":
          description: catalog-svc unreachable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
  /v1/packages/{id}:
    get:
      tags: [Catalog]
      summary: Active package detail
      operationId: getPackageById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Package ULID, `pkg_`-prefixed.
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetPackageResponse" }
        "404":
          description: Package not found or not publicly visible.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "500":
          description: Internal error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "502":
          description: catalog-svc unreachable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
  /v1/package-departures/{id}:
    get:
      tags: [Catalog]
      summary: Active package-departure detail
      operationId: getPackageDepartureById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Departure ULID, `dep_`-prefixed.
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetDepartureResponse" }
        "404":
          description: Departure not found or not publicly visible.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "500":
          description: Internal error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "502":
          description: catalog-svc unreachable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }

  # ---------------------------------------------------------------------------
  # Staff catalog write routes (BL-CAT-014 / S1-E-07)
  # All require Bearer token + catalog.package.manage permission.
  # Response shapes are the same read models (PackageDetail, DepartureDetail)
  # so the frontend can reuse its existing display components.
  # ---------------------------------------------------------------------------
  /v1/packages/create:
    # POST /v1/packages is handled inline under /v1/packages above, but since
    # oapi-codegen merges methods on the same path, we declare it here as a
    # separate path alias so it appears in the spec for documentation purposes.
    # The actual route registration is done manually in cmd/server.go.
    post:
      tags: [Catalog-Staff]
      summary: Create a catalog package (draft)
      operationId: createPackage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePackageBody" }
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetPackageResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "403":
          description: Missing catalog.package.manage permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "500":
          description: Internal error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
  /v1/packages/{id}/update:
    put:
      tags: [Catalog-Staff]
      summary: Update a catalog package (partial)
      operationId: updatePackageById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePackageBody" }
      responses:
        "200":
          description: Package updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetPackageResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "403":
          description: Missing catalog.package.manage permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "404":
          description: Package not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
  /v1/packages/{id}/delete:
    delete:
      tags: [Catalog-Staff]
      summary: Soft-delete a catalog package (archive)
      operationId: deletePackageById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Package archived
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "403":
          description: Missing catalog.package.manage permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "404":
          description: Package not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
  /v1/packages/{id}/departures/create:
    post:
      tags: [Catalog-Staff]
      summary: Create a departure under a package
      operationId: createDeparture
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Package ULID
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDepartureBody" }
      responses:
        "201":
          description: Departure created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetDepartureResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "403":
          description: Missing catalog.package.manage permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
        "404":
          description: Package not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
      description: |
        Single-point edge authentication per ADR 0009 / F1-W7. Clients send
        `Authorization: Bearer <access_token>` on any protected route.
        gateway-svc validates the token once against `iam-svc.ValidateToken`
        (gRPC); backends do NOT re-validate. Failure modes: missing/malformed
        → 401 UNAUTHORIZED; iam-svc rejects → 401 UNAUTHORIZED; iam-svc
        unreachable → 502 SERVICE_UNAVAILABLE (fail-closed).

        No route is currently tagged with this scheme — first consumer lands
        with BL-IAM-018 / S1-E-12 when iam auth REST routes move to gateway.
  schemas:
...
    PackageKind:
      type: string
      enum: [umrah_reguler, umrah_plus, hajj_furoda, hajj_khusus, badal, financial, retail]
...
    GetDepartureResponse:
      type: object
      required: [departure]
      properties:
        departure: { $ref: "#/components/schemas/DepartureDetail" }

    # ── Catalog: package-specific ──────────────────────────────────────
...
    PageMeta:
      type: object
      required: [has_more]
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more: { type: boolean }
    ListPackagesResponse:
      type: object
      required: [packages, page]
      properties:
        packages:
          type: array
          items: { $ref: "#/components/schemas/PackageListItem" }
        page: { $ref: "#/components/schemas/PageMeta" }
    GetPackageResponse:
      type: object
      required: [package]
      properties:
        package: { $ref: "#/components/schemas/PackageDetail" }

    # ── IAM auth schemas (BL-IAM-018 / S1-E-12) ─────────────────────────
    # All auth routes proxied from iam-svc REST → gateway REST → iam-svc gRPC.
...